carrito funcionando con bugs

+++ Wallpic/database/models/Purchase.js
@@ -8,7 +8,7 @@ module.exports = (sequelize, DataTypes) => {
     
     order_number: DataTypes.INTEGER,
     
-    total: DataTypes.INTEGER,
+    total: DataTypes.DECIMAL(6,2),
     
     user_id: DataTypes.INTEGER,
